eliminate keys.go and move store keys to keepers.go
